+++ crates/bevy_plugin/src/example_ui/ui_setup.rs
@@ -1,0 +1,44 @@
+use bevy::prelude::*;
+
+pub(crate) fn ui_setup_plugin(app: &mut App) {
+    app.add_system(setup.on_startup());
+}
+
+fn setup(mut commands: Commands) {
+    // root node
+    commands
+        .spawn(NodeBundle {
+            style: Style {
+                size: Size::width(Val::Percent(100.0)),
+                justify_content: JustifyContent::Center,
+                align_items: AlignItems::FlexEnd,
+                margin: UiRect::bottom(Val::Px(10.0)),
+                ..default()
+            },
+            ..default()
+        })
+        .with_children(|parent| {
+            parent
+                .spawn(NodeBundle {
+                    style: Style {
+                        size: Size::width(Val::Percent(70.0)),
+                        min_size: Size::height(Val::Px(100.0)),
+                        border: UiRect::all(Val::Px(20.0)),
+                        flex_direction: FlexDirection::ColumnReverse,
+                        ..default()
+                    },
+                    background_color: Color::rgb(0.4, 0.4, 1.0).into(),
+                    ..default()
+                })
+                .with_children(|parent| {
+                    parent.spawn(NodeBundle {
+                        style: Style {
+                            size: Size::new(Val::Percent(100.0), Val::Percent(100.0)),
+                            ..default()
+                        },
+                        background_color: Color::rgb(0.8, 0.8, 1.0).into(),
+                        ..default()
+                    });
+                });
+        });
+}
